feat(terminal): share the full keyboard with open terminal
Replace the shell Keys pad with the same character, modifier, symbol, and function-key component used by Agent tools. Preserve the measured keyboard footprint and the shell local-input gate.

refs #302
