+++ Sources/Heeler/Terminal/TerminalKeyboard.swift
@@ -174,9 +174,9 @@ enum TerminalFunctionKey: Int, CaseIterable, Hashable {
     }
 }
 
-/// Terminal controls and characters exposed by Composer's tools keyboard.
-/// These are explicit actions rather than authored text, so they bypass the
-/// draft while the Ghostty surface itself remains display-only.
+/// Terminal controls and characters exposed by the shared full keyboard.
+/// In Composer mode these explicit actions bypass the draft while the
+/// Ghostty surface itself remains display-only.
 enum AgentQuickKey: Hashable {
     case escape
     case tab
@@ -316,71 +316,6 @@ enum TerminalControlKey: Equatable, CaseIterable {
     case right
     case enter
 
-    /// Backspace takes the top row's right edge, where the iOS keyboard puts it
-    /// and where a thumb finds it without looking. It trades places with ⌃Z
-    /// rather than crowding in, so the rows stay evenly sized.
-    static let rows: [[Self]] = [
-        [.escape, .tab, .controlC, .controlD, .backspace],
-        [.home, .pageUp, .up, .pageDown, .end],
-        [.controlZ, .left, .down, .right, .enter],
-    ]
-
-    var title: String? {
-        switch self {
-        case .escape: "Esc"
-        case .tab: "Tab"
-        case .controlC: "⌃C"
-        case .controlD: "⌃D"
-        case .controlZ: "⌃Z"
-        case .home: "Home"
-        case .pageUp: "PgUp"
-        case .pageDown: "PgDn"
-        case .end: "End"
-        case .up, .backspace, .left, .down, .right, .enter: nil
-        }
-    }
-
-    var systemImageName: String? {
-        switch self {
-        case .up: "arrow.up"
-        case .backspace: "delete.left"
-        case .left: "arrow.left"
-        case .down: "arrow.down"
-        case .right: "arrow.right"
-        case .enter: "return"
-        default: nil
-        }
-    }
-
-    var accessibilityLabel: String {
-        switch self {
-        case .escape: "Escape"
-        case .tab: "Tab"
-        case .controlC: "Control C"
-        case .controlD: "Control D"
-        case .controlZ: "Control Z"
-        case .home: "Home"
-        case .pageUp: "Page Up"
-        case .up: "Up Arrow"
-        case .pageDown: "Page Down"
-        case .end: "End"
-        case .backspace: "Backspace"
-        case .left: "Left Arrow"
-        case .down: "Down Arrow"
-        case .right: "Right Arrow"
-        case .enter: "Enter"
-        }
-    }
-
-    var repeats: Bool {
-        switch self {
-        case .home, .pageUp, .up, .pageDown, .end, .backspace, .left, .down, .right:
-            true
-        case .escape, .tab, .controlC, .controlD, .controlZ, .enter:
-            false
-        }
-    }
-
     func bytes(applicationCursor: Bool) -> [UInt8] {
         bytes(applicationCursor: applicationCursor, modifiers: [])
     }
@@ -445,229 +380,6 @@ enum TerminalControlKey: Equatable, CaseIterable {
 final class TerminalSuppressedSoftKeyboardView: UIView {
     override var intrinsicContentSize: CGSize {
         CGSize(width: UIView.noIntrinsicMetric, height: 0)
-    }
-}
-
-/// The control-key pane of the Keys dock. It fills whatever space the dock's
-/// tab container gives it. Driven by a closure rather than the terminal view
-/// so the app-side dock can outlive any one terminal surface.
-final class TerminalControlPadView: UIView {
-    private let send: (TerminalControlKey) -> Void
-    private let toggleModifier: (TerminalKeyModifiers) -> Void
-    private var modifierButtons: [TerminalKeyModifiers: UIButton] = [:]
-
-    /// Mirrors `TerminalKeyboardControl.pendingModifiers`, pushed from
-    /// `TerminalControlPadRepresentable.updateUIView` so the ⌃/⌥ caps stay
-    /// highlighted while armed, including arms set from another surface.
-    var armedModifiers = TerminalKeyModifiers() {
-        didSet { guard armedModifiers != oldValue else { return }; refreshModifierButtons() }
-    }
-
-    init(
-        send: @escaping (TerminalControlKey) -> Void,
-        toggleModifier: @escaping (TerminalKeyModifiers) -> Void = { _ in }
-    ) {
-        self.send = send
-        self.toggleModifier = toggleModifier
-        super.init(frame: .zero)
-        configureKeys()
-    }
-
-    @available(*, unavailable)
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) is unavailable")
-    }
-
-    private func configureKeys() {
-        let rows = UIStackView()
-        rows.translatesAutoresizingMaskIntoConstraints = false
-        rows.axis = .vertical
-        rows.distribution = .fillEqually
-        rows.spacing = 8
-        addSubview(rows)
-
-        rows.addArrangedSubview(makeModifierRow())
-        for keys in TerminalControlKey.rows {
-            let row = UIStackView()
-            row.axis = .horizontal
-            row.distribution = .fillEqually
-            row.spacing = 8
-            for key in keys {
-                row.addArrangedSubview(makeButton(for: key))
-            }
-            rows.addArrangedSubview(row)
-        }
-
-        NSLayoutConstraint.activate([
-            rows.topAnchor.constraint(equalTo: topAnchor, constant: 10),
-            rows.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 10),
-            rows.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -10),
-            rows.bottomAnchor.constraint(equalTo: bottomAnchor, constant: -8),
-        ])
-    }
-
-    /// One-shot sticky modifiers: tap ⌃ or ⌥ to arm it for the next key only;
-    /// tapping the armed cap again disarms. Firing any key consumes the arm.
-    private func makeModifierRow() -> UIStackView {
-        let row = UIStackView()
-        row.axis = .horizontal
-        row.distribution = .fillEqually
-        row.spacing = 8
-        for modifier in [TerminalKeyModifiers.control, .option] {
-            let button = makeModifierButton(for: modifier)
-            modifierButtons[modifier] = button
-            row.addArrangedSubview(button)
-        }
-        return row
-    }
-
-    private func makeModifierButton(for modifier: TerminalKeyModifiers) -> UIButton {
-        let button = UIButton(type: .system)
-        button.configuration = modifierConfiguration(armed: false, modifier: modifier)
-        button.addAction(UIAction { [weak self] _ in
-            UIDevice.current.playInputClick()
-            self?.toggleModifier(modifier)
-        }, for: .touchUpInside)
-        button.accessibilityLabel =
-            modifier == .control ? "Control modifier" : "Option modifier"
-        button.accessibilityValue = "Not armed"
-        return button
-    }
-
-    private func refreshModifierButtons() {
-        for (modifier, button) in modifierButtons {
-            let armed = armedModifiers.contains(modifier)
-            button.configuration = modifierConfiguration(armed: armed, modifier: modifier)
-            button.accessibilityValue = armed ? "Armed" : "Not armed"
-        }
-    }
-
-    private func modifierConfiguration(
-        armed: Bool, modifier: TerminalKeyModifiers
-    ) -> UIButton.Configuration {
-        var configuration = armed ? UIButton.Configuration.filled() : UIButton.Configuration.gray()
-        configuration.title = modifier == .control ? "⌃" : "⌥"
-        configuration.baseForegroundColor = .label
-        if !armed {
-            configuration.baseBackgroundColor = .secondarySystemFill
-        }
-        configuration.cornerStyle = .medium
-        configuration.titleTextAttributesTransformer = UIConfigurationTextAttributesTransformer {
-            incoming in
-            var outgoing = incoming
-            outgoing.font = .preferredFont(forTextStyle: .body)
-            return outgoing
-        }
-        return configuration
-    }
-
-    private func makeButton(for key: TerminalControlKey) -> TerminalKeyButton {
-        var configuration = UIButton.Configuration.gray()
-        configuration.title = key.title
-        configuration.image = key.systemImageName.flatMap {
-            UIImage(systemName: $0, withConfiguration: UIImage.SymbolConfiguration(
-                textStyle: .body, scale: .medium))
-        }
-        configuration.baseForegroundColor = .label
-        configuration.baseBackgroundColor = .secondarySystemFill
-        configuration.cornerStyle = .medium
-        configuration.titleTextAttributesTransformer = UIConfigurationTextAttributesTransformer {
-            incoming in
-            var outgoing = incoming
-            outgoing.font = .preferredFont(forTextStyle: .body)
-            return outgoing
-        }
-
-        let button = TerminalKeyButton(configuration: configuration, repeats: key.repeats) {
-            [send] in
-            UIDevice.current.playInputClick()
-            send(key)
-        }
-        button.accessibilityLabel = key.accessibilityLabel
-        return button
-    }
-}
-
-/// A key fires when the finger lifts, not when it lands: the pad sits inside
-/// the pane pager, and a swipe that starts on a key must switch panes without
-/// also sending an Esc down the wire. Holding still repeats, so the arrows go
-/// on behaving like arrows.
-private final class TerminalKeyButton: UIButton {
-    private let keyAction: () -> Void
-    private let repeats: Bool
-    private var repeatDelayTimer: Timer?
-    private var repeatTimer: Timer?
-    /// A hold that has begun repeating already sent the key; letting go of it
-    /// must not send one more.
-    private var didRepeat = false
-
-    init(configuration: UIButton.Configuration, repeats: Bool, action: @escaping () -> Void) {
-        self.keyAction = action
-        self.repeats = repeats
-        super.init(frame: .zero)
-        self.configuration = configuration
-        isExclusiveTouch = true
-        addTarget(self, action: #selector(pressed), for: .touchDown)
-        addTarget(self, action: #selector(released), for: .touchUpInside)
-        addTarget(
-            self, action: #selector(abandoned),
-            for: [.touchUpOutside, .touchCancel, .touchDragExit])
-    }
-
-    @available(*, unavailable)
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) is unavailable")
-    }
-
-    override func didMoveToWindow() {
-        super.didMoveToWindow()
-        if window == nil {
-            cancelTimers()
-        }
-    }
-
-    @objc private func pressed() {
-        didRepeat = false
-        guard repeats else { return }
-
-        let timer = Timer(timeInterval: 0.45, target: self, selector: #selector(beginRepeating),
-                          userInfo: nil, repeats: false)
-        repeatDelayTimer = timer
-        RunLoop.main.add(timer, forMode: .common)
-    }
-
-    @objc private func beginRepeating() {
-        repeatDelayTimer = nil
-        didRepeat = true
-        keyAction()
-        let timer = Timer(timeInterval: 0.075, target: self, selector: #selector(repeatKey),
-                          userInfo: nil, repeats: true)
-        repeatTimer = timer
-        RunLoop.main.add(timer, forMode: .common)
-    }
-
-    @objc private func repeatKey() {
-        keyAction()
-    }
-
-    @objc private func released() {
-        if !didRepeat {
-            keyAction()
-        }
-        cancelTimers()
-    }
-
-    /// The finger left the key — dragged off it, or taken by the pager. Either
-    /// way the key was not pressed.
-    @objc private func abandoned() {
-        cancelTimers()
-    }
-
-    private func cancelTimers() {
-        repeatDelayTimer?.invalidate()
-        repeatDelayTimer = nil
-        repeatTimer?.invalidate()
-        repeatTimer = nil
     }
 }
