+++ Tests/HeelerTests/ShellTerminalStoreTests.swift
@@ -118,7 +118,7 @@ struct ShellTerminalStoreTests {
 
         #expect(terminal.isLocalInputEnabled)
         // No Snippets or Skills on a shell terminal: its Keys dock offers the
-        // control pad and Appearance alone.
+        // full keyboard and Appearance alone.
         #expect(ShellTerminalKeysDock.tabs == [.controls, .appearance])
         terminal.sendControlKey(.enter)
         try #require(
